build: ship a linux/armhf release, and make the tests compile for it
32-bit ARM is where the cheap hardware is, and madshare had no build for it.
`make release` now produces a .deb (armhf), an .rpm (armv7hl) and a tarball
for it alongside the existing four targets.

armhf is a PACKAGING name, not a GOARCH, and that is the whole shape of the
change: Go spells it GOARCH=arm plus a GOARM level, deb and rpm each spell it
differently again, and "arm" on its own would tell somebody downloading a
binary nothing about which of three incompatible ABIs it is for. So the loop
now carries the label separately from the toolchain values, and the FreeBSD
arm-less arms of it are named in the label too. ARMHF_GOARM picks the level:
7, which is what armhf means to Debian and Fedora, set explicitly because
that is also the toolchain default today and has not always been. A Pi 1 or
Zero is ARMv6 and needs ARMHF_GOARM=6.

One real 32-bit defect turned up, in a test: fillBytes multiplied by
0x9E3779B1 in int, which overflows a 32-bit int, so `go vet`/`go test` could
not COMPILE the federation package for GOARCH=arm — the whole suite would
have been unrunnable on the platform being shipped. Done in uint64 now, which
also leaves the byte pattern identical to what 64-bit hosts produced.

The other two usual 32-bit hazards were checked rather than assumed and are
absent: byte accounting is int64 throughout (the sole int conversion is a
chunk COUNT, which overflows past 512 TiB), and every 64-bit atomic is a
typed atomic.Int64, which cannot be misaligned the way a raw atomic.AddInt64
on a struct field can.

The 32-bit target is a full build, federation included. Verified under
qemu-arm-static, since this development host is Apple silicon and cannot
execute 32-bit ARM at all: SQLite, an 18 MB upload with tag extraction, eight
cover variants, ffprobe and fpcalc analysis, the yggdrasil transport, a
[[listen_mesh]] bind on the node's own mesh address, /api/madnetwork, and an
ordered shutdown — no panics. The other four targets were re-built and every
binary's architecture confirmed with file(1), because this change renamed
variables the FreeBSD packaging path also reads.

+++ federation/swarm_test.go
@@ -22,10 +22,17 @@ import (
 
 // fillBytes returns n deterministic, chunk-distinct bytes (so per-chunk hashes
 // differ and a multi-chunk transfer is a real test).
+//
+// The multiplier is 0x9E3779B1, which does not fit in a 32-bit int — so the
+// arithmetic is done in uint64 rather than in int, or this package fails to
+// COMPILE on a 32-bit platform (GOARCH=arm, i.e. the armhf release target) and
+// takes the whole test suite with it. uint64 also keeps the byte pattern
+// identical to what 64-bit hosts produced before, since i is never negative and
+// the products never came close to wrapping.
 func fillBytes(n int) []byte {
 	b := make([]byte, n)
 	for i := range b {
-		b[i] = byte((i*2654435761 + i/251) & 0xff)
+		b[i] = byte((uint64(i)*2654435761 + uint64(i)/251) & 0xff)
 	}
 	return b
 }
